multimodel-healthcare-rag-backend-master/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.17
httpx==0.28.1
groq==0.31.0
email-validator==2.2.0
aiofiles==24.1.0
PyPDF2==3.0.1
python-docx==1.1.2
